pyproject.toml
[tool.poetry]
name = "data-aware-dp"
version = "0.3.1"
description = ""
authors = ["Roy Rinberg <royrinberg@gmail.com>"]
readme = "README.md"
packages = [{include = "data_aware_dp"}]

[tool.poetry.scripts]
dadp-ml-training = "data_aware_dp.train_ml_cli:cli"


[tool.poetry.dependencies]
python = ">=3.9,<3.10"
matplotlib = "3.5.2"
pandas = "^1.4.3"
seaborn = "^0.11.2"
toolz = "^0.11.2"
tqdm = "^4.64.0"
plotly = "^5.9.0"
numpy = "^1.20.0, <=1.24.0"
numba = "^0.56.0"

torch = "^1.12.1"
torchvision = "^0.13.1"

PyYAML = "^6.0"
scikit-learn = "^1.1.2"
transformers = "^4.31.0"




[tool.poetry.group.dev.dependencies]

black = "^22.6.0"
yapf = "^0.32.0"
jupyter = "^1.0.0"
bpython = "^0.23"

# jax_privacy= { path = "jax_privacy/",develop = true}
# jaxlib = {url = "https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.1.73+cuda11.cudnn805-cp39-none-manylinux2010_x86_64.whl"}
# jaxline= {git = "https://github.com/deepmind/jaxline.git", branch = "master"}

#jax="0.3.23"
# opacus needs to be installed from DADP version

#opacus= { path = "opacus/",develop = true}
opacus = {path = "Opacus-PRV/", develop = true}

#jax = "0.3.23"
ipywidgets = "^7.0.0"
mpmath = "^1.2.1"
sympy = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
